==> notebooks/01_chembl.ipynb <==
# 01. ChEMBL Preprocessing

from pathlib import Path

import numpy as np
import pandas as pd
import sqlite3

Download 34th version of ChEMBL SQLite database from here https://ftp.ebi.ac.uk/pub/databases/chembl/ChEMBLdb/latest/ and unpack it:

!mkdir -p ../data
!wget https://ftp.ebi.ac.uk/pub/databases/chembl/ChEMBLdb/latest/chembl_34_sqlite.tar.gz -O ../data/chembl_34_sqlite.tar.gz
!tar -xzf ../data/chembl_34_sqlite.tar.gz -C ../data

DATA_DIR = Path('../data')
PATH_TO_CHEMBL = DATA_DIR / 'chembl_34/chembl_34_sqlite/chembl_34.db'

connection = sqlite3.connect(PATH_TO_CHEMBL)
chembl = connection.cursor()

request = '''
SELECT
    cs.canonical_smiles,
    cmpd.accession AS protein_accession,
    act.standard_type,
    act.standard_value,
    act.standard_units,
    act.standard_relation,
    td.organism,
    td.pref_name AS target_name,
    md.chembl_id AS compound_chembl_id,
    td.chembl_id AS target_chembl_id,
    a.assay_id,
    cmpd.sequence AS protein_sequence
FROM
    activities act
    JOIN assays a ON act.assay_id = a.assay_id
    JOIN target_dictionary td ON a.tid = td.tid
    JOIN target_components tc ON td.tid = tc.tid
    JOIN component_sequences cmpd ON tc.component_id = cmpd.component_id
    JOIN organism_class oc ON td.tax_id = oc.tax_id
    JOIN molecule_dictionary md ON act.molregno = md.molregno
    JOIN compound_structures cs ON md.molregno = cs.molregno
    JOIN docs d ON a.doc_id = d.doc_id
WHERE
    a.confidence_score = 9
    AND td.target_type = 'SINGLE PROTEIN'
    AND oc.L2 = 'Mammalia'
    AND act.standard_type IN ('Ki', 'Kd', 'IC50')
    AND act.standard_value IS NOT NULL
    AND act.standard_units IN ('nM')
    AND act.standard_flag = 1
    AND act.potential_duplicate = 0
    AND act.data_validity_comment IS NULL
    AND act.activity_comment IS NULL
    AND a.assay_type IN ('B', 'F')
    AND act.src_id NOT IN (SELECT src_id FROM source WHERE src_description LIKE '%unreliable_source%')
ORDER BY
    td.pref_name, act.standard_value;
'''

result = chembl.execute(request).fetchall()

data = pd.DataFrame(result, columns=[
    'smiles',
    'uniprot_id',
    'type',
    'value',
    'units',
    'relation',
    'organism',
    'target_name',
    'compound_chembl_id',
    'target_chembl_id',
    'assay_id',
    'sequence'
])
data

data['units'].value_counts()

For the further filtration we need less columns.

data = data[[
    'smiles',
    'type',
    'value',
    'units',
    'relation',
    'sequence',
    'uniprot_id',
]]
data

rows_to_drop = data['relation'].isin(['~', '>>'])
data = data[~rows_to_drop]

Create columns with value and sign for each data type

def get_activities(row):
    relation = row['relation']
    # Standardize the relation between datasets
    if relation == '>=':
        relation = '>'
    if relation == '<=':
        relation = '<'
    
    # If value or relation is missing, return NaN
    ki = np.nan
    ki_sign = None
    kd = np.nan
    kd_sign = None
    ic50 = np.nan
    ic50_sign = None
    ec50 = np.nan
    ec50_sign = None

    if row['type'] == 'Ki':
        ki = row['value']
        ki_sign = relation
    elif row['type'] == 'Kd':
        kd = row['value']
        kd_sign = relation
    elif row['type'] == 'IC50':
        ic50 = row['value']
        ic50_sign = relation
    elif row['type'] == 'EC50':
        ec50 = row['value']
        ec50_sign = relation
    
    return ki, kd, ic50, ec50, ki_sign, kd_sign, ic50_sign, ec50_sign

activities = pd.DataFrame(
    data.apply(get_activities, axis=1).to_list(),
    columns=['ki', 'kd', 'ic50', 'ec50', 'ki_sign', 'kd_sign', 'ic50_sign', 'ec50_sign']
)

data = pd.concat([
    data[['smiles', 'sequence', 'uniprot_id']], 
    activities
], axis=1)
data['source'] = 'chembl'
data

data.to_csv(DATA_DIR / 'chembl.csv', index=False)
