==> notebooks/02_bindingdb.ipynb <==
# 02. BindingDB Preprocessing

from pathlib import Path

import pandas as pd
import numpy as np

Download Nov 2024's version of BindingDB and unzip it: https://www.bindingdb.org/bind/downloads/BindingDB_All_202411_tsv.zip

!mkdir -p ../data
!wget https://www.bindingdb.org/bind/downloads/BindingDB_All_202411_tsv.zip -O ../data/BindingDB_All_202411_tsv.zip
!unzip -o ../data/BindingDB_All_202411_tsv.zip -d ../data

DATA_DIR = Path('../data')
PATH_TO_BDB = DATA_DIR / 'BindingDB_All.tsv'

bdb = pd.read_csv(PATH_TO_BDB, sep='\t', on_bad_lines='skip', low_memory=False, usecols=[
    'Ligand SMILES',
    'Ki (nM)',
    'IC50 (nM)',
    'Kd (nM)',
    'EC50 (nM)',
    'UniProt (SwissProt) Primary ID of Target Chain',
    'UniProt (TrEMBL) Primary ID of Target Chain',
    'BindingDB Target Chain Sequence',
    'Number of Protein Chains in Target (>1 implies a multichain complex)',
])

bdb = bdb[bdb['Number of Protein Chains in Target (>1 implies a multichain complex)'] == 1]

bdb = bdb.rename({
    'Ligand SMILES': 'smiles',
    'Ki (nM)' : 'ki',
    'IC50 (nM)': 'ic50',
    'Kd (nM)': 'kd',
    'EC50 (nM)': 'ec50',
    'UniProt (SwissProt) Primary ID of Target Chain': 'uniprot_swissprot',
    'UniProt (TrEMBL) Primary ID of Target Chain': 'uniprot_trembl',
    'BindingDB Target Chain Sequence': 'sequence',
},  axis=1) 
bdb

def extract_sign(val):
    if isinstance(val, float):
        return None  # it's nan
    val = str(val)
    if not val:
        return None
    if val[0] in ["<", ">"]:
        return val[0]
    else:
        return '='

def extract_val(val):
    if isinstance(val, float):
        return None  # it's nan
    val = str(val)
    if not val:
        return None
    if val[0] in ["<", ">"]:
        return float(val[1:])
    else:
        return float(val)

bdb['ki_sign'] = bdb['ki'].map(extract_sign)
bdb['ki'] = bdb['ki'].map(extract_val)
bdb['ic50_sign'] = bdb['ic50'].map(extract_sign)
bdb['ic50'] = bdb['ic50'].map(extract_val)
bdb['kd_sign'] = bdb['kd'].map(extract_sign)
bdb['kd'] = bdb['kd'].map(extract_val)
bdb['ec50_sign'] = bdb['ec50'].map(extract_sign)
bdb['ec50'] = bdb['ec50'].map(extract_val)

def get_uniprot_id(row):
    if pd.notna(row['uniprot_swissprot']):
        return row['uniprot_swissprot']
    else:
        return row['uniprot_trembl']

bdb = bdb.dropna(subset=['uniprot_swissprot', 'uniprot_trembl'], how='all').reset_index(drop=True)
bdb['uniprot_id'] = bdb.apply(get_uniprot_id, axis=1)
bdb = bdb.drop(labels=[
    'uniprot_swissprot',
    'uniprot_trembl',
    'Number of Protein Chains in Target (>1 implies a multichain complex)'
], axis=1).reset_index(drop=True)
bdb['source'] = 'bdb'
bdb

bdb.to_csv(DATA_DIR / 'bdb.csv', index=False)
