==> notebooks/02_biolip.ipynb <==
# 02. BioLip Preprocessing

import pandas as pd
import numpy as np
from gemmi import cif

Download BioLip2 here: https://zhanggroup.org/BioLiP/download/BioLiP.txt.gz

PATH_TO_BIOLIP = '../../mol-finder/data/raw/biolip/BioLiP.txt.gz'
biolip = pd.read_csv(PATH_TO_BIOLIP, compression='gzip', sep='\t', header=None)
biolip

Instead of having SMILES of the ligands, BioLip contains ligand ids in the Chemical Component Dictionary used by PDB database. We need to convert them back to SMILES.

Download the CCD: https://files.wwpdb.org/pub/pdb/data/monomers/components.cif

PATH_TO_CCD = '../../mol-finder/data/raw/biolip/components.cif'
ccd = cif.read(PATH_TO_CCD)
ccd

# Parse SMILES from CCD
ligand_ids = list(set(biolip[4].to_list()))
ligand_smiles = []
for ligand_id in ligand_ids:
    try:
        ligand = ccd[ligand_id]
    except:
        ligand_smiles.append(None)
        continue
    descriptor_types = list(ligand.find_values('_pdbx_chem_comp_descriptor.type'))
    is_smiles_found = False
    for i, dtype in enumerate(descriptor_types):
        if dtype == 'SMILES':
            ligand_smiles.append(ligand.find_values('_pdbx_chem_comp_descriptor.descriptor')[i].strip('"'))
            is_smiles_found = True
            break
    if not is_smiles_found:
        ligand_smiles.append(None)

ligand_to_smiles = dict(zip(ligand_ids, ligand_smiles))
result_smiles = biolip[4].map(ligand_to_smiles)

Now let's process the activities

result = biolip[[0, 13, 14, 15, 16, 17, 20]]
result.columns = [
    'PDB',
    'activity_lit',
    'activity_bmoad',
    'activity_pdbbind',
    'activity_bdb',
    'uniprot_id',
    'sequence',
]
result['smiles'] = result_smiles
result

lit = ~result['activity_lit'].isna()
moad = ~result['activity_bmoad'].isna()
pdbbind = ~result['activity_pdbbind'].isna()
bdb = ~result['activity_bdb'].isna()
has_activity = lit | moad | pdbbind | bdb

result_with_actives = result[has_activity]
result_with_actives 

# We don't need BDB data, because we will join it later
no_bdb = result_with_actives['activity_bdb'].isna()
result_no_bdb = result_with_actives[no_bdb]
result_no_bdb

We need to parse the activities to work with them

import re

def parse_value(line):
    if type(line) is float:
        return None
    line = line.strip()
    if not line or line.lower() == 'nan':
        return None
    # Split the line by commas
    segments = line.split(',')
    for segment in segments:
        segment = segment.strip()
        # Ignore any log values
        if segment.startswith(('Ki=', 'Kd=', 'IC50=')):
            # Use regex to extract the measurement
            match = re.match(r'(Ki|Kd|IC50)=([<>~]=?)?([0-9.]+)([a-zA-Zμ]+M)$', segment)
            if match:
                measure_type = match.group(1)
                inequality = match.group(2) or ''
                if inequality.strip('=') in ('<', '>', '<=', '>=', '~'):
                    return None  # Contains inequality sign
                value = float(match.group(3))
                unit = match.group(4)
                # Convert units to nM
                if unit in ('nM', 'nM'):
                    value_in_nM = value
                elif unit in ('uM', 'μM'):
                    value_in_nM = value * 1000
                elif unit == 'pM':
                    value_in_nM = value * 0.001
                else:
                    return None  # Unsupported unit
                return (measure_type, value_in_nM)
    return None  # No Ki, Kd, or IC50 value found


result_no_bdb['activity_pdbbind_parsed'] = result_no_bdb['activity_pdbbind'].map(parse_value)
result_no_bdb['activity_bmoad_parsed'] = result_no_bdb['activity_bmoad'].map(parse_value)
result_no_bdb['activity_lit_parsed'] = result_no_bdb['activity_lit'].map(parse_value)

non_empty_new_data = result_no_bdb.dropna(subset=['activity_pdbbind_parsed', 'activity_bmoad_parsed', 'activity_lit_parsed'], how='all').reset_index(drop=True)
non_empty_new_data

def get_activity(row):
    if row['activity_pdbbind_parsed'] is not None:
        return row['activity_pdbbind_parsed']
    if row['activity_bmoad_parsed'] is not None:
        return row['activity_bmoad_parsed']
    return row['activity_lit_parsed']

def unfold_activities(value):
    # Return (Kd, Ki, IC50)
    measure, activity = value
    if measure == 'Kd':
        return (activity, None, None)
    if measure == 'Ki':
        return (None, activity, None)
    if measure == 'IC50':
        return (None, None, activity)

single_activity = non_empty_new_data.apply(get_activity, axis=1)
single_activity

unfolded_activities = single_activity.apply(unfold_activities)
unfolded_activities = pd.DataFrame(unfolded_activities.to_list(), columns=['Kd', 'Ki', 'IC50'])
unfolded_activities

biolip = pd.DataFrame({
    'smiles': non_empty_new_data['smiles'],
    'uniprot_id': non_empty_new_data['uniprot_id'],
    'sequence': non_empty_new_data['sequence'],
    'kd': unfolded_activities['Kd'],
    'ki': unfolded_activities['Ki'],
    'ic50': unfolded_activities['IC50'],
})
biolip['source'] = 'biolip'
biolip['ec50'] = None
biolip['ec50_sign'] = None
biolip['ki_sign'] = biolip['ki'].apply(lambda x: None if pd.isna(x) else '=')
biolip['kd_sign'] = biolip['kd'].apply(lambda x: None if pd.isna(x) else '=')
biolip['ic50_sign'] = biolip['ic50'].apply(lambda x: None if pd.isna(x) else '=')
biolip

biolip.to_csv('biolip.csv', index=False)
