==> notebooks/04_merge.ipynb <==
# 04. Merge the datasets

import multiprocessing as mp
from pathlib import Path

import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
from rdkit.rdBase import BlockLogs
from tqdm import tqdm

from standardize import standardize

DATA_DIR = Path('../data/')

block = BlockLogs()

bdb = pd.read_csv(DATA_DIR / 'bdb.csv')
biolip = pd.read_csv(DATA_DIR / 'biolip.csv')
chembl = pd.read_csv(DATA_DIR / 'chembl.csv')

data = pd.concat([bdb, biolip, chembl], ignore_index=True)
data = data.dropna(subset=['ki', 'kd', 'ic50', 'ec50'], how='all')
data = data.dropna(subset=['smiles', 'sequence', 'uniprot_id'], how='any')
data

with mp.Pool(mp.cpu_count()) as pool:
    data['smiles'] = list(
        tqdm(pool.imap(standardize, data['smiles']), total=len(data), desc="Standardizing SMILES")
    )

# Remove rows with invalid SMILES after standardization
data = data[~data['smiles'].isna()].reset_index(drop=True)
# Filter out SMILES with asterisks
data = data[~data['smiles'].str.contains('\*', na=False)].reset_index(drop=True)
data

# Prefer BindingDB data over BioLip and ChEMBL
group_has_bdb = data.groupby(['smiles', 'sequence'])['source'].transform(lambda x: x.eq('bdb').any())
mask = (~group_has_bdb) | (data['source'] == 'bdb')
data = data[mask].reset_index(drop=True)

data

data.to_csv(DATA_DIR / 'merged_activities.csv', index=False)

# Get PLINDER split

data = pd.read_csv(DATA_DIR / 'merged_activities.csv', low_memory=False)

PATH_TO_PLINDER_ANNOTATION = "gs://plinder/2024-06/v2/index/annotation_table.parquet"
PATH_TO_PLINDER_SPLIT = "gs://plinder/2024-06/v2/splits/split.parquet"

annotation_data = pd.read_parquet(PATH_TO_PLINDER_ANNOTATION)
annotation_data

split_data = pd.read_parquet(PATH_TO_PLINDER_SPLIT)
split_data

We split activity data into train/val/test based on proteins that are assigned to systems that are split by PLINDER:

# Map system_id to uniprot_id from PLINDER annotation
system2uniprot_id = annotation_data.set_index('system_id')['system_pocket_UniProt'].to_dict()
split_data['uniprot_id'] = split_data['system_id'].map(system2uniprot_id)
# Find split which certain uniprot_id belongs to. Sinse some UniProt IDs appear in multiple sets, we will prioritize 
# the validation set, the training set, and finally the test set to have a larger validation set, 
# and don't include those proteins in the test set.
split_to_idx = {'train': 1, 'val': 0, 'test': 2}
split_data['split_idx'] = split_data['split'].map(split_to_idx)
split_data = split_data.sort_values('split_idx').drop_duplicates('uniprot_id', keep='first')
uniprot_id2split = split_data.set_index('uniprot_id')['split'].to_dict()
data['split'] = data['uniprot_id'].map(uniprot_id2split)

data['split'].value_counts()

# Remove datapoints belonging to the 'removed' split or 
# aren't in the intersection of the merged dataset and the PLINDER dataset
to_remove = (data['split'] == 'removed') | data['split'].isna()
data = data[~to_remove].reset_index(drop=True)
data

data.to_csv(DATA_DIR / 'merged_activities_split.csv', index=False)

# PAINS filtering

data = pd.read_csv(DATA_DIR / 'merged_activities_split.csv', low_memory=False)

from pains_filter import filter_pains

data = filter_pains(data, 'smiles')

data.to_csv(DATA_DIR / 'merged_activities_split_nopains.csv', index=False)

# Binarization and deduplication

data = pd.read_csv(DATA_DIR / 'merged_activities_split_nopains.csv', low_memory=False)
data

Take Ki and Kd values only for validation and test sets as these values are the most accurate, and IC50 has to be ambigously scaled (see [paper](https://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0061007&type=printable)).

train = data[data['split'] == 'train'].reset_index(drop=True)
val = data[data['split'] == 'val'].reset_index(drop=True)
test = data[data['split'] == 'test'].reset_index(drop=True)

is_ki_nan = test['ki'].isna()
is_kd_nan = test['kd'].isna()
val = val[~val['ki'].isna() | ~val['kd'].isna()]
test = test[~test['ki'].isna() | ~test['kd'].isna()]
print(f"Val set size after taking Ki and Kd values only: {len(val)}")
print(f"Test set size after taking Ki and Kd values only: {len(test)}")

def binarize_row(row, threshold=1000):
    activity_cols = ['ki', 'kd']
    sign_cols = ['ki_sign', 'kd_sign']
    for activity_col, sign_col in zip(activity_cols, sign_cols):
        value = row[activity_col]
        sign = row[sign_col]
        if pd.notnull(value):
            # We want the activity values to be exclusively compared to the threshold, i.e val < threshold.
            # If the value is given with an equality sign, it can be directly compared to the threshold.
            # For the cases with inequality signs, the values are typically given as > or <, and we need to
            # check if the value is inclusively less or greater than the threshold (e.g. <10000 nM should have 
            # an "active" binary flag).
            if sign == '=':
                return 1 if value < threshold else 0
            elif sign == '>':
                return 0 if value >= threshold else np.nan
            elif sign == '<':
                return 1 if value <= threshold else np.nan
            else:
                return np.nan
    
    return np.nan  # No activity value found

val['is_active'] = val.apply(lambda row: binarize_row(row, threshold=1000), axis=1)
test['is_active'] = test.apply(lambda row: binarize_row(row, threshold=1000), axis=1)

The binary activity label must be consistent for a given pair of (SMILES, sequence). Here is where inconsistency deduplication comes into play:

def is_consistent(group):
    labels = group['is_active'].dropna().unique()
    return len(labels) == 1

# Filter groups with consistent labels
val_consistent_groups = val.groupby(['smiles', 'sequence']).filter(is_consistent)
test_consistent_groups = test.groupby(['smiles', 'sequence']).filter(is_consistent)

# Deduplicate by keeping one row per group
val = val_consistent_groups.groupby(['smiles', 'sequence'], as_index=False).first()
test = test_consistent_groups.groupby(['smiles', 'sequence'], as_index=False).first()

print(f"Val set size after inconsistency deduplication: {len(val)}")
print(f"Test set size after inconsistency deduplication: {len(test)}")

val['is_active'].value_counts()

test['is_active'].value_counts()

val = val.drop(columns=['ki', 'kd', 'ki_sign', 'kd_sign', 'ic50', 'ec50', 'ic50_sign', 'ec50_sign'])
test = test.drop(columns=['ki', 'kd', 'ki_sign', 'kd_sign', 'ic50', 'ec50', 'ic50_sign', 'ec50_sign'])

val

test

We leave train set as is for people to be able to work with all activity values including Ki, Kd, IC50 and EC50 along with all inequality signs and inconsistencies thus providing a larger training set if this helps their models:

train

PLUMBER_DIR = DATA_DIR / 'plumber'
PLUMBER_DIR.mkdir(exist_ok=True, parents=True)

val.to_csv(PLUMBER_DIR / 'val.csv', index=False)
test.to_csv(PLUMBER_DIR / 'test.csv', index=False)
train.to_csv(PLUMBER_DIR / 'train.csv', index=False)
